requirements.txt
torch>=2.1.0
transformers>=4.46.0
trl>=0.12.0
datasets>=2.16.0
accelerate>=0.25.0
pyyaml>=6.0
